checkstyle | lombok(Getter, Setter) 추가
