--- src/main/java/com/spirnt/mission/discodeit/service/UserService.java
@@ -1,7 +1,0 @@
-package com.spirnt.mission.discodeit.service;
-
-import com.spirnt.mission.discodeit.entity.User;
-
-public interface UserService extends CRUDService<User>{
-
-}
