+++ src/main/java/com/sprint/mission/discodeit/service/UserService.java
@@ -1,0 +1,7 @@
+package com.sprint.mission.discodeit.service;
+
+import com.sprint.mission.discodeit.entity.User;
+
+public interface UserService extends CRUDService<User>{
+
+}
